Bug fix-North version
Now it won't make duplicate entries of the same object when program is relaunched. 

NOTE: This one is set up for Pears North.  control F for north and replace with south. Also replace the _n_ with _s_ . Finally uncomment the sci = for the candles south and comment out the current one for north.